requirements.txt
anthropic
google-genai
pillow
python-dotenv
requests
